Project 8 - Task 5 - commit 5

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -8,7 +8,7 @@
 
 int _sqrt_recursion(int n)
 {
-	unsigned int i = n;
+	long i = n;
 
 	if (i < 0)
 		return (-1);
